Added more tests

+++ src/main/java/com/espen/ws/web/api/OfferController.java
@@ -15,12 +15,16 @@ import org.springframework.web.bind.annotation.RequestMethod;
 
 import com.espen.ws.model.Offer;
 import com.espen.ws.services.OffersServiceInterface;
+import com.espen.ws.services.UsersServiceInterface;
 
 @Controller
 public class OfferController {
 
 	@Autowired
 	private OffersServiceInterface offersService;
+	
+	@Autowired
+	private UsersServiceInterface usersService;
 
 	@RequestMapping(value = "/api/offers", method = RequestMethod.POST, consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
 	public ResponseEntity<Offer> createOffer(@Validated @RequestBody Offer offer) {
@@ -38,11 +42,18 @@ public class OfferController {
 	@RequestMapping(value = "/api/offers/{id}", method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE)
 	public ResponseEntity<Offer> getOffersById(@PathVariable("id") Integer id) {
 		Offer offer = offersService.findOneById(id);
+		if(offer == null){
+			return new ResponseEntity<Offer>(HttpStatus.NOT_FOUND);
+		}
 		return new ResponseEntity<Offer>(offer, HttpStatus.OK);
 	}
 
 	@RequestMapping(value = "/api/offers/username/{username}", method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE)
 	public ResponseEntity<Collection<Offer>> getOffersByUsername(@PathVariable("username") String username) {
+	
+		if( usersService.findOne(username) == null ){
+			return new ResponseEntity<Collection<Offer>>(HttpStatus.NOT_FOUND);
+		}
 		Collection<Offer> offers = offersService.findByUsername(username);
 		return new ResponseEntity<Collection<Offer>>(offers, HttpStatus.OK);
 	}
@@ -50,12 +61,18 @@ public class OfferController {
 	@RequestMapping(value = "/api/offers/{id}", method = RequestMethod.PUT, consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
 	public ResponseEntity<Offer> updateOffer(@PathVariable Integer id, @RequestBody Offer offer) {
 		offer.setId(id);
+		if(offersService.findOneById(id)== null){
+			return new ResponseEntity<Offer>(HttpStatus.NOT_FOUND);
+		}
 		Offer updatedOffer = offersService.update(offer);
 		return new ResponseEntity<Offer>(updatedOffer, HttpStatus.OK);
 	}
 
 	@RequestMapping(value = "/api/offers/{id}", method = RequestMethod.DELETE)
 	public ResponseEntity<Offer> deleteOffer(@PathVariable Integer id) {
+		if(offersService.findOneById(id)== null){
+			return new ResponseEntity<Offer>(HttpStatus.NOT_FOUND);
+		}
 		Offer offerToDelete = offersService.findOneById(id);
 		offersService.delete(offerToDelete);
 		return new ResponseEntity<Offer>(HttpStatus.NO_CONTENT);
